refactor: used seperate service file for chat
